Fix: Incorrect way of finding five high straights
Previously I was just finding the total value of the cards but
there can be a situation where the total value of the cards is same
as that in five high straights such as in "9D 3S 2C 7S 7C" but this
is a "one pair" hand and the function (_is_five_high_straight) was
returning this as "true".

Fixing this with comparing the card values directly. Example:
"9D 3S 2C 7S 7C" -> [2, 3, 7, 7, 9] != [2, 3, 4, 5, 14]

Even in the rare case of giving this multiple times to the function,
the function will return "True" and in other times it won't change
the position of ace (14) to the start in the _cards attribute.